UserService refactored and new controller
Some methods from UserService were renamed
Method checkPassword was moved to LoginController
